# perf(h5p-types): decrease tsc instantiations