--- a/packages/h5p-types/src/types/InferL10nType.ts
+++ b/packages/h5p-types/src/types/InferL10nType.ts
@@ -1,11 +1,14 @@
 import type { ReadonlyDeep } from "type-fest";
 import type { H5PFieldGroup } from "./H5PField";
 
+export type L10nGroup = ReadonlyDeep<H5PFieldGroup & { name: "l10n" }>;
+
+type FieldNames<TField extends L10nGroup> = TField["fields"][number]["name"];
+
 export type InferL10nType<
   TField extends ReadonlyDeep<H5PFieldGroup & { name: "l10n" }>,
-  TFieldNames extends string = TField["fields"][number]["name"],
 > = {
   l10n: {
-    [TKey in TFieldNames]: string;
+    [TKey in FieldNames<TField>]: string;
   };
 };