--- a/packages/h5p-types/src/types/InferParamsFromSemantics.ts
+++ b/packages/h5p-types/src/types/InferParamsFromSemantics.ts
@@ -1,6 +1,6 @@
 import type { ReadonlyDeep } from "type-fest";
-import type { H5PField, H5PFieldGroup } from "./H5PField";
-import type { InferL10nType } from "./InferL10nType";
+import type { H5PField } from "./H5PField";
+import type { InferL10nType, L10nGroup } from "./InferL10nType";
 import type { InferOptionalWithDefault } from "./InferParamTypeFromFieldType";
 
 /**
@@ -41,13 +41,12 @@ import type { InferOptionalWithDefault } from "./InferParamTypeFromFieldType";
  */
 export type InferParamsFromSemantics<
   TSemantics extends ReadonlyArray<ReadonlyDeep<H5PField>>,
-> = TSemantics extends readonly [infer TField, ...infer TRestFields]
-  ? TField extends ReadonlyDeep<H5PField>
-    ? TRestFields extends ReadonlyArray<ReadonlyDeep<H5PField>>
-      ? (TField extends ReadonlyDeep<H5PFieldGroup & { name: "l10n" }>
-          ? InferL10nType<TField>
-          : Record<TField["name"], InferOptionalWithDefault<TField>>) &
-          InferParamsFromSemantics<TRestFields>
-      : unknown
-    : unknown
+> = TSemantics extends readonly [
+  infer TField extends ReadonlyDeep<H5PField>,
+  ...infer TRestFields extends ReadonlyArray<ReadonlyDeep<H5PField>>,
+]
+  ? (TField extends ReadonlyDeep<L10nGroup>
+      ? InferL10nType<TField>
+      : Record<TField["name"], InferOptionalWithDefault<TField>>) &
+      InferParamsFromSemantics<TRestFields>
   : unknown;