## 📝 Description

<!--
Please provide a brief description of the purpose and context of this
pull request. What problem does it address or what feature does it
introduce? This will help reviewers understand the changes at a high
level. Include any relevant background information or references to
issues/feature requests.
-->

TS caches by type names. By creating more utility types, more operations can be cached.

From:

```
Files:                         412
Lines of Library:            39791
Lines of Definitions:       144414
Lines of TypeScript:          5417
Lines of JavaScript:             0
Lines of JSON:                   0
Lines of Other:                  0
Identifiers:                196437
Symbols:                    142447
Types:                       11184
Instantiations:             238459
Memory used:               218006K
Assignability cache size:    14643
Identity cache size:            48
Subtype cache size:              1
Strict subtype cache size:       0
I/O Read time:               0.13s
Parse time:                  0.34s
ResolveModule time:          0.02s
ResolveTypeReference time:   0.01s
ResolveLibrary time:         0.01s
Program time:                0.55s
Bind time:                   0.18s
Check time:                  0.32s
printTime time:              0.00s
Emit time:                   0.00s
Total time:                  1.05s
```

To:

```
Files:                         412
Lines of Library:            39791
Lines of Definitions:       144414
Lines of TypeScript:          5423
Lines of JavaScript:             0
Lines of JSON:                   0
Lines of Other:                  0
Identifiers:                196449
Symbols:                    136904
Types:                        9363
Instantiations:             149178
Memory used:               210717K
Assignability cache size:    18910
Identity cache size:            24
Subtype cache size:              1
Strict subtype cache size:       0
I/O Read time:               0.03s
Parse time:                  0.33s
ResolveModule time:          0.02s
ResolveTypeReference time:   0.01s
ResolveLibrary time:         0.01s
Program time:                0.45s
Bind time:                   0.18s
Check time:                  0.30s
printTime time:              0.00s
Emit time:                   0.00s
Total time:                  0.93s
```


## ✅ Checklist

<!--
Please check all that apply. Any checkboxes that don't apply to your
pull request can be removed.
-->

- [x] **Pull Request Title and Description**: Is the pull request title concise and descriptive? Does the description provide enough context for reviewers to understand your changes?
- [x] **Code Quality**: Have you ensured that your code adheres to the project's coding standards and conventions?
- [x] **Review**: Have you gone through your own code changes to catch any potential issues before submitting for review?
- [x] **Testing**: Have you added appropriate tests to cover the changes you've made? Are existing tests passing?

<!--
Thank you for your contribution to h5p-types! Your attention to
detail and following of these guidelines will greatly assist the
review process and help maintain the project's quality and consistency.
-->